Define an explicit main entry point for the tests executable
The main entry point for gtest based executables is defined in gtest_main
for convenience. However, The cares static library which is built as part
of the grpc dependency contains multiple "main" entry points as well.
The linker (at least in Linux) simply chooses the first main in encounters
which happened to be the one provided by "cares".

By defining an explicit main, we ensure it's the first "main" symbol the
linker sees and then everything is hunky dory.