QemuVirtualMachine: Use VM name for cloud-init instance-id.
This is a workaround for the fact that “ubuntu launch” doesn't actually launch a
new VM at the moment; it runs r/w against the downloaded image directly.

This means that the cloud-init config is only applied if the instance-id is different,
so make sure the instance-id is different to make testing not a flaming pile of weird.