exec/connect: assemble ssh command on the client
For linux platforms, ask the daemon for the vm port and private ssh key.
Assemble the ssh command args from the client.

The private ssh key blob is saved to a file by the client process to avoid
file permission issues.

Since the ssh command is now assembled on the client, remove all
VirtualMachineExecute related artifacts.